refactor: move default.ts, and react.ts in init directory
- introduce a directory structure of commandName/[template1, template2]
- move default and react template generators in init/*
- avoid namespace issues and cleaner structure
