+++ packages/create-webpack-app/src/plopfile.ts
@@ -1,6 +1,6 @@
 import { NodePlopAPI } from "./types";
 
 export default async function (plop: NodePlopAPI) {
-  await plop.load("./generators/default.js", {}, true);
-  await plop.load("./generators/react.js", {}, true);
+  await plop.load("./generators/init/default.js", {}, true);
+  await plop.load("./generators/init/react.js", {}, true);
 }
